requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
dataclasses-json==0.5.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.51.0
django-environ==0.9.0
django-extensions==3.2.0
django-fcm==0.1.1
django-filter==22.1
django-rest-auth==0.9.5
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
facebook-sdk==3.1.0
factory-boy==3.2.1
Faker==14.2.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-facebook-api==0.14.5
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.1
types-cryptography==3.3.23
typing-extensions==4.3.0
typing-inspect==0.8.0
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
